// src/adapters/ravepage/api-client/models/TrackLinkCreateIn.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TrackLinkCreateIn = {
    description?: string;
    link_type?: string;
    platform?: 'spotify' | 'apple_music' | 'soundcloud' | 'youtube' | 'bandcamp' | 'beatport' | 'tidal' | 'deezer' | 'audius' | 'other';
    title?: string;
    url?: string;
};