// src/adapters/ravepage/api-client/models/TrackLinkOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TrackLinkOut = {
    created_at?: string;
    description?: string;
    id?: string;
    link_type?: string;
    platform?: 'youtube' | 'soundcloud' | 'bluesky' | 'twitch' | 'x' | 'instagram' | 'spotify';
    title?: string;
    track_id?: string;
    updated_at?: string;
    url?: string;
};