// src/adapters/ravepage/api-client/models/TrackLinkUpdateIn.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TrackLinkUpdateIn = {
    description?: string;
    link_type?: string;
    platform?: 'spotify' | 'apple_music' | 'soundcloud' | 'youtube' | 'bandcamp' | 'beatport' | 'tidal' | 'deezer' | 'audius' | 'other';
    title?: string;
    url?: string;
};